+++ resale/src/main/java/com/goti/resale/infra/dto/TicketGameInfo.java
@@ -1,4 +1,4 @@
-package com.goti.resale.dto.internal;
+package com.goti.resale.infra.dto;
 
 import java.util.UUID;
